+++ components/index.js
@@ -1,0 +1,10 @@
+export {default as Article} from './Article';
+export {default as Block} from './Block';
+export {default as Button} from './Button';
+export {default as Form} from './Form';
+export {default as Grid} from './Grid';
+export {default as Navigator} from './Navigator';
+export {default as NewsList} from './NewsList';
+export {default as Panel} from './Panel';
+export {default as Subject} from './Subject';
+export {default as Typography} from './Typography';